// projects/projects/src/api/model/scenario.ts
/**
 * MaestroHost - SceneService
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { LearningPath } from './learningPath';
import { Scene } from './scene';

export interface Scenario {
  scenarioId?: number;
  name?: string;
  description?: string;
  numberOfVisits?: number;
  numberOfLearningPath?: number;
  createdAtUTC?: Date;
  inUse?: boolean;
  scene?: Scene;
  learningPaths?: Array<LearningPath>;
  readonly sceneId?: number;
}